feat: Implement multi-profile CV system with print styling
- Add profile selection landing page with audience-based routing
- Implement toolbar with dark mode toggle (persists preference)
- Add centered breadcrumbs with navigation links
- Create grouped dropdown menu (profiles → collapsible audiences)
- Add print styles: two-column layout, inline skills, centered highlights
- Update skills component with spacing between categories
- Archive old data files to src/data-old/
- Remove obsolete ProfileSelectionPage component
- Set Myriad Pro as global font family
