+++ src/data-old/contact_details.json
@@ -1,0 +1,13 @@
+{
+  "firstName": "Yomi",
+  "lastName": "Colledge",
+  "email": "yomi@boodah.net",
+  "location": "UK REMOTE",
+  "phone": "07894 987 855",
+  "linkedin": "https://www.linkedin.com/in/yomicolledge/",
+  "github": "https://github.com/baphled",
+  "portfolio": {
+    "link": "https://n-vyro.io/",
+    "link": "https://github.com/boodah-consulting/quik-cv"
+  }
+}
