+++ src/data-old/recruiters.yaml
@@ -1,0 +1,170 @@
+---
+first_name: Yomi
+last_name: Colledge
+email: yomi@boodah.net
+location: Remote (UK)
+phone: +44 7894 987 855
+links:
+  - label: LinkedIn
+    url: https://www.linkedin.com/in/yomicolledge
+  - label: GitHub
+    url: http://github.com/baphled
+  - label: Portfolio
+    url: http://boodah.net
+summary: |
+  **Senior Ruby Engineer / Lead Developer / Principal Technical Consultant**
+
+  Senior engineer with 15+ years’ experience delivering mission-critical Ruby systems across startups, scale-ups, and government platforms.
+
+highlights: |
+  Acting as Principal Engineer and technical lead for an early-stage startup, owning architecture, delivery, and technical strategy.
+
+jobs:
+  - company: "FullSpektrum®"
+    position: "Principal Technical Consultant"
+    start_date: "Jul 2024"
+    end_date: "Nov 2025"
+    description: |
+      - Defined and executed end-to-end technical strategy and delivery roadmap for a startup, ensuring alignment with business objectives
+      - Advised on technology selection and architecture to showcase the company’s core USP effectively
+      - Designed MVP development and demonstration strategy to accelerate time-to-market
+      - Prepared comprehensive technical due-diligence materials, supporting successful investor engagement
+      - Developed and implemented a strategy for beta version development, optimising product feedback and iteration
+
+  - company: "Mindful Chef"
+    position: "Senior Ruby Developer"
+    start_date: "Mar 2022"
+    end_date: "Apr 2024"
+    description: |
+      - Engineered platform migration to Ruby 3.1, reducing server load and improving system performance
+      - Designed and implemented innovative delivery flow, driving **30% increase in weekly delivery volume**
+      - Re-architected complex Kafka-based pipelines, resulting in:
+        - **45% reduction in delivery-box debt collection**
+        - Eliminated payment failures
+        - Minimised data loss risks
+      - Automated critical workflows, generating **£132,000 annual operational cost savings**
+      - Delivered cross-functional features spanning ordering, subscriptions, logistics, and CMS
+
+  - company: "Spice Rack"
+    position: "Ruby DevOps Consultant"
+    start_date: "Jan 2022"
+    end_date: "Mar 2022"
+    description: |
+      - Rapidly developed AWS-backed provisioning and content delivery services using Ruby + Docker
+      - Compressed project timeline from 6 to 3 months while maintaining high-quality delivery
+      - Implemented robust backup + rollback system, reducing infrastructure setup time by **80%**
+      - Delivered production-ready infrastructure with minimal client overhead
+      - Enabled seamless handover to internal teams for ongoing maintenance
+      - Ensured high availability and scalability for anticipated traffic spikes
+
+  - company: "Freelance Technical Consultant"
+    position: "Senior Ruby Developer & Architect"
+    start_date: "Oct 2020"
+    end_date: "Dec 2021"
+    description: |
+      - Architected scale-able white-label SaaS platforms and high-performance APIs
+      - Implemented ELK stack for real-time monitoring of multi-million event streams
+      - Developed specialised tools including pension calculation systems and advanced publishing workflows
+      - Designed REST APIs that enhanced user on-boarding and reduced client churn
+
+  - company: "NTTData / BEIS / Defra / MAS"
+    position: "Senior Ruby Development Consultant"
+    start_date: "Jun 2019"
+    end_date: "Oct 2020"
+    description: |
+      - Led migration of critical government platforms, ensuring reliability, compliance, and accessibility
+      - Standardised Ruby development practices across multiple teams
+      - Integrated complex REST and SOAP APIs for inter-agency data exchange
+      - Stabilised large-scale public-sector services under high-traffic conditions
+      - Rapidly integrated with cross-functional team to deliver strategic digital tools
+      - Introduced defra-ruby-style guide (https://github.com/DEFRA/defra-ruby-style), enhancing code quality and consistency across teams
+      - Implemented Dough open-source framework (https://github.com/moneyadviceservice/dough)
+      - Developed and deployed Pensions Calculator Tool
+      - Maintained internal CMS with high efficiency, consistently delivering ahead of schedule
+
+  - company: "Diverse Technical Consulting Portfolio"
+    position: "Senior Ruby Developer & Technical Architect"
+    start_date: "Jun 2012"
+    end_date: "Aug 2018"
+    description: |
+      - Established ELK stack solutions for advanced data analysis and operational insights
+      - Architected flexible white-label platforms and high-performance APIs
+      - Managed complex migrations, integrations, and infrastructure scaling across varied technology stacks
+      - Demonstrated adaptability across multiple industries and technological challenges
+      - Contributed to HSBC Global Connections changes and enhanced the Atlas project.
+projects:
+  - name: "n-vyro.io"
+    description: |
+      Architected a comprehensive IoT ecosystem demonstrating end-to-end technical leadership.
+      Developed custom hardware and software solutions, implementing CI/CD across 14+ repositories
+      to ensure reliable, scalable product delivery. Showcased ability to manage complex,
+      multi-domain technical projects from prototype to production.
+    url: "https://n-vyro.io"
+    key_achievements:
+      - Managed 14+ interconnected repositories
+      - Implemented robust CI/CD pipelines
+      - Bridged hardware and software development domains
+
+  - name: "QuikCV"
+    description: |
+      Open-source CV generation tool highlighting technical versatility and commitment to
+      best practices. Migrated from Ruby to Vue.js, demonstrating technology adaptation
+      and architectural evolution. Serves as a living portfolio of clean architecture
+      and modern development practices.
+    url: "https://github.com/boodah-consulting/quik-cv"
+    key_achievements:
+      - Successfully migrated between technology stacks
+      - Maintained open-source project as a technical showcase
+      - Demonstrated CLI tooling and architectural best practices
+skills:
+  Languages:
+    - Ruby
+    - Go
+    - C/C++
+    - JavaScript / ECMAScript
+
+  Frameworks:
+    - Ruby on Rails
+    - Sinatra
+    - Grape
+    - React
+    - Vue.js
+    - Quasar Framework
+
+  "Data & Messaging":
+    - Kafka
+    - PostgreSQL
+    - MySQL
+    - Redis
+    - ELK Stack
+    - GraphQL
+    - MongoDB
+    - Elasticsearch
+
+  "Cloud & DevOps":
+    - AWS (EC2, S3, RDS, CloudFront)
+    - Docker
+    - CI/CD (Jenkins, Travis CI, GitHub Actions)
+    - Semantic Versioning
+    - Capistrano
+    - Chef
+    - Ansible
+    - Logstash / Kibana
+    - StatsD / Graphite
+
+  "Engineering Practices":
+    - Agile / Scrum / Kanban
+    - TDD / BDD
+    - API Design
+    - Object-Oriented Programming
+    - Component Driven Development
+    - Pair Programming
+    - Technical Mentoring
+    - Strategic Consulting
+    - Observability & Performance Optimisation
+    - Platform Modernisation
+    - Rescue Missions
+    - Cost Reduction Strategies
+    - Release Management
+    - Infrastructure as Code
+    - Cross-functional Collaboration
